main: refactor to run defer function. handle non-default -f. handle compiler

+++ main.go
@@ -33,18 +33,31 @@ clean:
 ## EOF
 `
 
-var g_verbose = flag.Bool("v", false, "enable verbose output")
-var g_makefile = flag.String("f", ".makefile-mkgo.mk", "alternate Makefile")
-var g_race = flag.Bool("race", false, "enable build with race detector")
+const mkfname = ".makefile-mkgo.mk"
+
+var (
+	g_verbose  = flag.Bool("v", false, "enable verbose output")
+	g_makefile = flag.String("f", mkfname, "alternate Makefile")
+	g_race     = flag.Bool("race", false, "enable build with race detector")
+	g_compiler = flag.String("compiler", "gc", "go compiler to use (gc,gccgo,llgo)")
+)
 
 func main() {
 	flag.Parse()
 
-	err := ioutil.WriteFile(*g_makefile, []byte(tmpl), 0644)
-	defer os.Remove(*g_makefile)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "**error** creating file [%s]: %v\n", *g_makefile, err)
-		os.Exit(1)
+	rc := run()
+	os.Exit(rc)
+}
+
+func run() int {
+	var err error
+	if *g_makefile == mkfname {
+		err = ioutil.WriteFile(*g_makefile, []byte(tmpl), 0644)
+		defer os.Remove(*g_makefile)
+		if err != nil {
+			fmt.Fprintf(os.Stderr, "**error** creating file [%s]: %v\n", *g_makefile, err)
+			return 1
+		}
 	}
 
 	cmdargs := make([]string, 0, flag.NArg())
@@ -56,6 +69,8 @@ func main() {
 	if *g_race {
 		goflags += " -race"
 	}
+	goflags += fmt.Sprintf(" -compiler=%s", *g_compiler)
+
 	os.Setenv("GOFLAGS", goflags)
 
 	cmdargs = append(cmdargs, flag.Args()...)
@@ -70,7 +85,8 @@ func main() {
 			strings.Join(cmdargs, " "),
 			err,
 		)
-		os.Exit(1)
+		return 1
 	}
 
+	return 0
 }
